+++ byline-feed/playwright.config.js
@@ -8,7 +8,7 @@ module.exports = defineConfig( {
 		timeout: 10_000,
 	},
 	use: {
-		baseURL: process.env.WP_BASE_URL || 'http://localhost:8886',
+		baseURL: process.env.WP_BASE_URL || 'http://localhost:8896',
 		headless: process.env.PLAYWRIGHT_HEADLESS !== '0',
 		trace: 'retain-on-failure',
 		screenshot: 'only-on-failure',